requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
email==6.0.0a1
Pillow==8.1.0
psycopg2==2.8.6
python-http-client==3.3.1
pytz==2020.5
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
